Fix the build on FreeBSD i386
* Fix errors about unused function for sha2_is_supported

* libspl needs -Wno-error=atomic-alignment for i386

* Printf format specifier.
  See https://github.com/freebsd/freebsd-src/commit/9198651515b

* Fix 32-bit build with GCC.
  See https://github.com/freebsd/freebsd-src/commit/2d76470b701

Sponsored by:   ConnectWise

+++ include/sys/zio.h
@@ -82,7 +82,8 @@ gbh_nblkptrs(uint64_t size) {
 static inline zio_eck_t *
 gbh_eck(zio_gbh_phys_t *gbh, uint64_t size) {
 	ASSERT(IS_P2ALIGNED(size, sizeof (blkptr_t)));
-	return ((zio_eck_t *)((uintptr_t)gbh + size - sizeof (zio_eck_t)));
+	return ((zio_eck_t *)((uintptr_t)gbh + (size_t)size -
+	    sizeof (zio_eck_t)));
 }
 
 static inline blkptr_t *
